[add] PaLMApiDataSourceの単体テストを整備
